+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 # Change Log
 
+
+## UNRELEASED
+
+### Added
+
+- Shutdown handler